Arrumado CategoriaDAO
Resolvido problema com metodo listarTudo

+++ siaproDesktop/src/siapro/dao/CategoriaDAO.java
@@ -86,11 +86,11 @@ public class CategoriaDAO implements InterfaceDAO {
 	}
 
 	@Override
-	public Entidade pesquisarId(int id) {
+	public Entidade pesquisarId(long id) {
 		String sql = "SELECT * FROM categoria WHERE id = ?";
 		try {
 			stmt = conexao.prepareStatement(sql);
-			stmt.setInt(1, id);
+			stmt.setLong(1, id);
 			ResultSet rs = stmt.executeQuery();
 			
 			if (rs.next()) {
